+++ daemon/graphdriver/zfs/zfs_linux.go
@@ -5,19 +5,20 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/graphdriver"
+	"github.com/docker/docker/daemon/internal/fstype"
 )
 
 func checkRootdirFs(rootDir string) error {
-	fsMagic, err := graphdriver.GetFSMagic(rootDir)
+	fsMagic, err := fstype.GetFSMagic(rootDir)
 	if err != nil {
 		return err
 	}
 	backingFS := "unknown"
-	if fsName, ok := graphdriver.FsNames[fsMagic]; ok {
+	if fsName, ok := fstype.FsNames[fsMagic]; ok {
 		backingFS = fsName
 	}
 
-	if fsMagic != graphdriver.FsMagicZfs {
+	if fsMagic != fstype.FsMagicZfs {
 		log.G(context.TODO()).WithField("root", rootDir).WithField("backingFS", backingFS).WithField("storage-driver", "zfs").Error("No zfs dataset found for root")
 		return graphdriver.ErrPrerequisites
 	}
